Refactor : 리팩포링 및 주석 수정
