+++ package.json
@@ -14,13 +14,9 @@
   "dependencies": {
     "@hookform/resolvers": "^4.1.3",
     "@tailwindcss/oxide": "^4.0.17",
-    "axios": "^1.8.3",
-    "classnames": "^2.5.1",
-    "cookie": "^1.0.2",
     "date-fns": "^4.1.0",
     "framer-motion": "^12.5.0",
     "google-auth-library": "^9.15.1",
-    "js-cookie": "^3.0.5",
     "lightningcss": "^1.29.3",
     "lottie-react": "^2.4.1",
     "moment": "^2.30.1",
@@ -29,7 +25,6 @@
     "next-auth": "^5.0.0-beta.25",
     "react": "^19.0.0",
     "react-calendar": "^5.1.0",
-    "react-cookie": "^8.0.1",
     "react-dom": "^19.0.0",
     "react-hook-form": "^7.54.2",
     "react-toastify": "^11.0.5",
@@ -71,7 +66,6 @@
     "postcss": "^8.5.3",
     "prettier": "^3.5.3",
     "prettier-plugin-tailwindcss": "^0.6.11",
-    "storybook": "^8.6.4",
     "tailwindcss": "^4.0.12",
     "ts-node": "^10.9.2",
     "typescript": "^5",
