+++ src/app/sample/_components/Jin.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import LatestCommentSection from '@/app/main/_conponents/LatestCommentSection';
+import LatestCommentSection from '@/app/main/_components/LatestCommentSection';
 import EpigramCommentSection from '@/app/feed/[id]/_component/EpigramCommentSection';
 import MyCommentList from '@/app/mypage/_components/MyCommentList';
 import { SessionProvider } from 'next-auth/react';
