+++ src/middleware.ts
@@ -1,45 +1,33 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { cookies } from 'next/headers';
-import { getToken } from 'next-auth/jwt'; // getToken을 사용하여 JWT 토큰 확인
+import { getToken } from 'next-auth/jwt';
 
 export const middleware = async (request: NextRequest) => {
-  // 쿠키에서 accessToken과 refreshToken을 읽기
-  const cookieStore = await cookies();
-  const accessToken = cookieStore.get('accessToken');
-  const refreshToken = cookieStore.get('refreshToken');
-
-  // JWT 토큰을 가져오기 위한 getToken 사용
   const token = await getToken({
     req: request,
     secret: process.env.NEXTAUTH_SECRET,
     secureCookie: process.env.NEXTAUTH_URL?.startsWith('https://'),
   });
-  const isAuthenticated = !!(accessToken?.value || refreshToken?.value || token); // 쿠키 또는 JWT 토큰으로 인증 여부 확인
+  const isAuthenticated = !!token;
 
   const pathname = request.nextUrl.pathname;
 
-  // 로그인 전 접근 가능한 경로 (로그인 전 경로에 대해서만 리디렉션을 처리)
-  const isBeforeLoginRoute = pathname === '/auth/login' || pathname === '/auth/signup' || pathname === '/';
+  const beforeLogin = pathname === '/auth/login' || pathname === '/auth/signup' || pathname === '/';
+  //로그인 전 접근 가능 경로로
 
-  // 로그인 후 접근 가능한 경로 (로그인 후 경로에 대해서만 리디렉션을 처리)
-  const isAfterLoginRoute =
-    pathname.startsWith('/mypage') || pathname.startsWith('/main') || pathname.startsWith('/feed');
+  // 로그인 후 접근 가능한 경로
+  const afterLogin = pathname.startsWith('/mypage') || pathname.startsWith('/main') || pathname.startsWith('/feed');
 
-  // 로그인 전 경로에 로그인된 상태로 접근 시 홈 페이지로 리디렉션
-  if (isBeforeLoginRoute && isAuthenticated) {
-    return NextResponse.redirect(new URL('/main', request.nextUrl)); // 홈 페이지로 리디렉션
+  if (beforeLogin && isAuthenticated) {
+    return NextResponse.redirect(new URL('/main', request.nextUrl));
   }
 
-  // 로그인 후 경로에 로그인되지 않은 상태로 접근 시 로그인 페이지로 리디렉션
-  if (isAfterLoginRoute && !isAuthenticated) {
-    return NextResponse.redirect(new URL('/auth/login', request.nextUrl)); // 로그인 페이지로 리디렉션
+  if (afterLogin && !isAuthenticated) {
+    return NextResponse.redirect(new URL('/auth/login', request.nextUrl));
   }
 
-  // 기본적으로 리디렉션 없이 진행
   return NextResponse.next();
 };
 
-// matcher 설정: 모든 경로에 대해 middleware가 동작하도록 설정
 export const config = {
   matcher: ['/login', '/signup', '/mypage', '/auth/login', '/auth/signup', '/main', '/feed', '/:path*'],
 };
